Ensure env gcc is used instead of system level one
* downgrade mkl and fix compilers

MKL 2025 is not compatible with that version of pytorch, which has been
compiled against an earlier version of it.
`compilers` ensures that the gcc version installed is actually also
accessible and the environment does not default to the system version.

* chore: pin channel for torch to pytorch

If not pinned, it can occur that the solver chooses conda-forge as
source instead, effectively installing the cpu only version. In this
case the kernels cannot be compiled anymore then.